Move fixExcelFilterZeroes function to scripts.js
