+++ test/blocks/search-results/search-results.test.js
@@ -221,37 +221,4 @@ describe('Search Results', () => {
     const pageWidget = block.children[4];
     expect(pageWidget.className.toString()).to.equal('pagination');
   });
-
-  it('Fix Excel Filter Zeroes', () => {
-    const data = [
-      {
-        unrelated: 0,
-        more_unrel: '0',
-        description: '0',
-        breadcrumbtitle: '0',
-        newsdate: '0',
-      },
-      {
-        unrelated: 0,
-        more_unrel: '0',
-        description: '0 Mydesc 0',
-        breadcrumbtitle: '',
-        newsdate: '12345',
-      },
-    ];
-
-    scripts.fixExcelFilterZeroes(data);
-
-    expect(data[0].unrelated).to.equal(0);
-    expect(data[0].more_unrel).to.equal('0');
-    expect(data[0].description).to.equal('');
-    expect(data[0].breadcrumbtitle).to.equal('');
-    expect(data[0].newsdate).to.equal('');
-
-    expect(data[1].unrelated).to.equal(0);
-    expect(data[1].more_unrel).to.equal('0');
-    expect(data[1].description).to.equal('0 Mydesc 0');
-    expect(data[1].breadcrumbtitle).to.equal('');
-    expect(data[1].newsdate).to.equal('12345');
-  });
 });
